add encode handler library - separate the logic from DBTool
