+++ app/ipcSocket/tool/CMakeLists.txt
@@ -14,6 +14,7 @@ add_executable(${PROJECT_NAME} ${SRC_FILES})
 target_include_directories(${PROJECT_NAME} PRIVATE
     include
     ../lib/fileHandler/include
+    ../lib/encodeHandler/include
 )
 
 # Link with OpenSSL
@@ -21,7 +22,9 @@ find_package(OpenSSL REQUIRED)
 target_link_libraries(${PROJECT_NAME} OpenSSL::SSL OpenSSL::Crypto)
 
 # Link with fileHandlerLib
-target_link_libraries(${PROJECT_NAME} "${CMAKE_SOURCE_DIR}/../lib/build/libBuild/libfileHandlerLib.so")
+target_link_libraries(${PROJECT_NAME} 
+    "${CMAKE_SOURCE_DIR}/../lib/build/libBuild/libencodeHandlerLib.so"
+    "${CMAKE_SOURCE_DIR}/../lib/build/libBuild/libfileHandlerLib.so")
 
 # Specify any other compiler flags or options as needed
 # set(CMAKE_C_FLAGS "${CMAKE_C_FLAGS} -Wall -Wextra")
